fix: reconstruct the config setting
